feat(packaging): update project metadata and settings
Remove '-dirty' suffix from subproject commit identifier to indicate a clean state.
Change project name in package.json to a lowercase and hyphenated format for consistency.
Improve readability of the description in package.json by adding a space between "Polymorphic" and "Serializations".
Add <IsPackable>false</IsPackable> to DeserializeFileMessages.csproj to specify that the project is not intended for NuGet packaging.